gofmt to make the linter happy
